change: add some visual tweaks, fix schema, alter type of component for child allowance
