+++ services/ui-src/src/measures/2021/CommonQuestions/types.ts
@@ -1,3 +1,5 @@
+import { OmsDataNode } from "./OptionalMeasureStrat";
+
 export interface DefinitionOfPopulation {
   DefinitionOfDenominator: string[];
   "DefinitionOfDenominator-Other": string;
@@ -94,11 +96,6 @@ export interface DeviationFields {
   other: string;
 }
 
-interface AggregateRate {
-  subRate?: RateFields[];
-  total?: RateFields[];
-}
-
 export interface OtherRatesFields {
   description?: string;
   rate?: RateFields[];
@@ -160,71 +157,14 @@ export interface PerformanceMeasureDescriptions {
 }
 
 export interface OptionalMeasureStratification {
-  //OptionalMeasureStratification
-  CategoriesReported: string[];
-
-  AddtnlEthnicity: string[];
-  AddtnlEthnicityRates: AggregateRate[];
-
-  AddtnlNonHispanicRace: string[];
-  AddtnlNonHispanicRaceRates: AggregateRate[];
-  AddtnlNonHispanicRaceSubCatTitle: { titles: string[] }[];
-  AddtnlNonHispanicRaceSubCatOptions: string[][];
-  AddtnlNonHispanicRaceSubCatRates: { rates: AggregateRate[] }[];
-
-  AddtnlNonHispanicSubCat: string[];
-  AddtnlNonHispanicSubCatRates: AggregateRate[];
-
-  NonHispanicRacialCategories: string[];
-  "NHRC-WhiteRates": AggregateRate;
-  "NHRC-BlackOrAfricanAmericanRates": AggregateRate;
-  "NHRC-AmericanIndianOrAlaskaNativeRates": AggregateRate;
-  "NHRC-AggregateAsianRates": AggregateRate;
-  "NHRC-IndependentAsianRates": AggregateRate[];
-  "NHRC-AggregateHawaiianOrPacificIslanderRates": AggregateRate;
-  "NHRC-IndependentHawaiianOrPacificIslanderRates": AggregateRate[];
-
-  EthnicityCategories: string[];
-  EthnicitySubCategories: string[];
-  NonHispanicEthnicityRates: AggregateRate;
-  HispanicIndependentReporting: string;
-  HispanicEthnicityAggregateRate: AggregateRate;
-  IndependentHispanicOptions: string[];
-  IndependentHispanicRates: AggregateRate[];
-
-  AsianIndependentReporting: string;
-  IndependentAsianOptions: string[];
-  NativeHawaiianIndependentReporting: string;
-  IndependentNativeHawaiianOptions: string[];
-
-  SexOptions: string[];
-  MaleSexRates: AggregateRate;
-  FemaleSexRates: AggregateRate;
-
-  PrimaryLanguageOptions: string[];
-  AddtnlPrimaryLanguage: string[];
-  AddtnlPrimaryLanguageRates: AggregateRate[];
-  EnglishLanguageRate: AggregateRate;
-  SpanishLanguageRate: AggregateRate;
-
-  DisabilityStatusOptions: string[];
-  DisabilitySSIRate: AggregateRate;
-  DisabilityNonSSIRate: AggregateRate;
-  AddtnlDisabilityStatusDesc: string;
-  AddtnlDisabilityRate: AggregateRate;
-
-  GeographyOptions: string[];
-  UrbanGeographyRate: AggregateRate;
-  RuralGeographyRate: AggregateRate;
-  AddtnlGeographyDesc: string;
-  AddtnlGeographyRate: AggregateRate;
-
-  ACAGroupRate: AggregateRate;
-
   OptionalMeasureStratification: {
     options: string[]; //checkbox
     selections: {
       [option: string]: TopLevelOmsNode;
     };
   };
+}
+
+export namespace DataDrivenTypes {
+  export type OptionalMeasureStrat = OmsDataNode[];
 }
