+++ services/ui-src/src/measures/2021/IETAD/index.tsx
@@ -194,6 +194,7 @@ export const IETAD = ({
               performanceMeasureDescriptions={
                 PerformanceMeasureData.performanceMeasureDescriptions
               }
+              adultMeasure
             />
           )}
         </>
